requirements.txt
streamlit
mysql-connector-python
python-dotenv  # (Optional, if using environment variables)